V2 adds 2nd timezone support & sweep secondhand
Tidied up config page, added 2nd timezone to sub-dial options.  Added battery-killer option for 5 refreshes of the second hand per second

+++ src/c/utils/MathUtils.c
@@ -33,9 +33,15 @@ GPoint add_points(GPoint a, GPoint b) {
 }
 
 // Converts an angle and a distance to a cartesian point.
+// GPoint polar_to_point(int angle, int distance) {
+//   int x = distance * ((double)cos_lookup(DEG_TO_TRIGANGLE(angle)) / ((double)TRIG_MAX_ANGLE));
+//   int y = distance * ((double)sin_lookup(DEG_TO_TRIGANGLE(angle)) / ((double)TRIG_MAX_ANGLE));
+//   return GPoint(x, y);
+// }
+
 GPoint polar_to_point(int angle, int distance) {
-  int x = distance * ((double)cos_lookup(DEG_TO_TRIGANGLE(angle)) / ((double)TRIG_MAX_ANGLE));
-  int y = distance * ((double)sin_lookup(DEG_TO_TRIGANGLE(angle)) / ((double)TRIG_MAX_ANGLE));
+  int x = (distance * cos_lookup(DEG_TO_TRIGANGLE(angle))) / TRIG_MAX_ANGLE;
+  int y = (distance * sin_lookup(DEG_TO_TRIGANGLE(angle))) / TRIG_MAX_ANGLE;
   return GPoint(x, y);
 }
